requirements.txt
python-dotenv
aiogram
google.generativeai
python-dotenv